test: isolate and speed up slow sandbox tests
