+++ .changeset/isolated-vercel-tests.md
@@ -1,0 +1,6 @@
+---
+'@openai/agents-core': patch
+'@openai/agents-extensions': patch
+---
+
+test: isolate and speed up slow sandbox tests
